tests/justhtml-tests/head_noscript_recovery.dat

#data
<head><noscript><!DOCTYPE html></noscript>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,31): unexpected-doctype

#script-off
#document
| <html>
|   <head>
|     <noscript>
|   <body>

#data
<head><noscript></p></noscript>
#errors
(1,1): expected-doctype-but-got-start-tag
(1,20): unexpected-end-tag

#script-off
#document
| <html>
|   <head>
|     <noscript>
|   <body>
